Initial commit: Generated by OPS-X
Build a web application called 'MILLION DOLLAR IDEA'. Pages needed: Home, Dashboard Requirements: I want a super crisp web app that is more like on the gothic side of the UI with a Noir theme and then the focus ogf the web app should be to generat billion dollar ideas. And make sure it works 100%!!, Let's get this bread dawggggg